requirements.txt
coverage
nose
py-bcrypt
pylint
pyramid
pyramid-debugtoolbar
pyramid-deform
pyramid-tm
SQLAlchemy
waitress
zope.sqlalchemy
